requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.3
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
